update to float type + compile with keys

+++ semestr1/prg5/fun.h
@@ -1,3 +1,3 @@
 #pragma once
-int Input(const char *filename, int **array,int *size);
+int Input(const char *filename, float **array,int *size);
 int Sdvig(float *array, int size);
